+++ spb/backends/bokeh/renderers/zgrid.py
@@ -5,7 +5,7 @@ def _draw_zgrid_helper(renderer, data):
     p, s = renderer.plot, renderer.series
     xi_dict, wn_dict, tp_dict, ts_dict = data
 
-    lkw = p.grid_line_kw
+    lkw = {"line_color": "#aaa", "line_dash": "dotted"}
     kw = p.merge({}, lkw, s.rendering_kw)
 
     def _add_labels(x, y, labels):